// src/app/users/schemas/business.schema.ts
import { Type } from "@sinclair/typebox";

export const businessSchema = {
  create: Type.Object({
    name: Type.String(),
    email: Type.String(),
    phone: Type.String(),
    whatsappNumber: Type.Optional(Type.String()),
    country: Type.String(),
    address: Type.String(),
    state: Type.String(),
    reference: Type.String(),
    industryId: Type.String(),
    bio: Type.Optional(Type.String()),
    cac: Type.Optional(Type.String()),
    logo: Type.Optional(Type.String()),
    govId: Type.Optional(Type.String()),
    cover: Type.Optional(Type.String()),
    slug: Type.Optional(Type.String()),
    twitterLink: Type.Optional(Type.String()),
    facebookLink: Type.Optional(Type.String()),
    instagramLink: Type.Optional(Type.String()),
    linkedinLink: Type.Optional(Type.String()),
  }),
  update: Type.Object({
    name: Type.Optional(Type.String()),
    email: Type.Optional(Type.String()),
    phone: Type.Optional(Type.String()),
    whatsappNumber: Type.Optional(Type.String()),
    country: Type.Optional(Type.String()),
    address: Type.Optional(Type.String()),
    state: Type.Optional(Type.String()),
    industry: Type.Optional(Type.String()),
    bio: Type.Optional(Type.String()),
    cover: Type.Optional(Type.String()),
    logo: Type.Optional(Type.String()),
    slug: Type.Optional(Type.String()),
    twitterLink: Type.Optional(Type.String()),
    facebookLink: Type.Optional(Type.String()),
    instagramLink: Type.Optional(Type.String()),
    linkedinLink: Type.Optional(Type.String()),
  }),
  delete: Type.Object({ id: Type.Number() }),
  get: Type.Object({
    id: Type.Number(),
    name: Type.String(),
    email: Type.String(),
    phone: Type.String(),
    whatsappNumber: Type.Optional(Type.String()),
    country: Type.Optional(Type.String()),
    address: Type.String(),
    state: Type.String(),
    industry: Type.Optional(Type.Any()),
    bio: Type.Optional(Type.String()),
    cac: Type.String(),
    cover: Type.String(),
    logo: Type.String(),
    slug: Type.String(),
    twitterLink: Type.Optional(Type.String()),
    facebookLink: Type.Optional(Type.String()),
    instagramLink: Type.Optional(Type.String()),
    linkedinLink: Type.Optional(Type.String()),
    products: Type.Optional(Type.Array(Type.Any())),

    createdAt: Type.String({ format: "date-time" }),
    updatedAt: Type.String({ format: "date-time" }),
  }),
};